tools/requirements.txt
scapy>=2.5.0
PyYAML>=6.0.2
requests>=2.32.3
pytest>=8.3.4
